+++ app/src/main/java/de/passwordvault/view/fragments/PackagesSelectedFragment.java
@@ -14,6 +14,7 @@ import de.passwordvault.R;
 import de.passwordvault.model.packages.Package;
 import de.passwordvault.view.utils.OnRecyclerItemClickListener;
 import de.passwordvault.view.utils.adapters.PackagesRecyclerViewAdapter;
+import de.passwordvault.view.utils.components.PasswordVaultBaseFragment;
 import de.passwordvault.viewmodel.activities.PackagesViewModel;
 
 
@@ -24,7 +25,7 @@ import de.passwordvault.viewmodel.activities.PackagesViewModel;
  * @author  Christian-2003
  * @version 3.5.0
  */
-public class PackagesSelectedFragment extends Fragment implements OnRecyclerItemClickListener<Package> {
+public class PackagesSelectedFragment extends PasswordVaultBaseFragment implements OnRecyclerItemClickListener<Package> {
 
     /**
      * Attribute stores the view model of the fragment and it's hosting activity.