+++ app/src/main/java/de/passwordvault/view/utils/components/PasswordVaultBaseActivity.java
@@ -1,0 +1,14 @@
+package de.passwordvault.view.utils.components;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+
+/**
+ * Clas implements the base activity for all activities within the app.
+ *
+ * @author  Christian-2003
+ * @version 3.5.1
+ */
+public abstract class PasswordVaultBaseActivity extends AppCompatActivity {
+
+}