+++ app/src/main/java/de/passwordvault/view/utils/components/PasswordVaultBaseFragment.java
@@ -1,0 +1,8 @@
+package de.passwordvault.view.utils.components;
+
+import androidx.fragment.app.Fragment;
+
+
+public abstract class PasswordVaultBaseFragment extends Fragment {
+
+}